We should be using a sql db instead of a text file for history
This massively simplifies the process, and includes me moving the gnarly
method within method coding_repl to extracted, more cohesive methods.
Especially because the LLM sucks ass at writing code with this amount of
complexity. A funny point to make, sure it took me 2h to refactor this
and think it through, but with no tests. An LLM just stomps all over it,
80% right. Idk about LLMs still lol. When long term complexity grows the
liability of them grow as well, much more exponentially than even a
junior development in my opinion.

Anyways, this lays the ground work for storing chats in a db instead,
and then, to start, dynamically generating the llm prompt on every
request, using the latest files. There are some kinks to work out, like
auto compacting in the case the developer updates a file after the LLM
writes a new version of it (and it is stored in a chat message), and the
latest version is in fact the wrong version (perhaps can bring back the
"look for max version" code).