We should ideally be taking advantage of native turn vocabularies
Since most models are trained natively to understand user and model
roles, using that, in this case via the types.UserContent and
types.ModelContent with python-genai, should make the conversational
aspect of this even better, more robust, and more performant.

In this case we use the system instructions for the static instructions,
then we add in the chat history after as a series of turns, then we add
the dynamic "database" context along with the user's query, or if
continuing, the last user's query, or if it was the model, being extra
explicit and giving it an anditional note to continue, as per its
instructions if so.